Add special error code for packets that come in over the uplink.
This can happen when the user has multiple tunnels and is routing packets sourced from the tunnel address over a different route toward the PoP, the packet thus arrives at the PoP on it's 'uplink' interface and the PoP sees that as a wrongly sourced packet.

This error logs that packet into the error for the tunnel that the source address belongs to.