Simplify another thing, while making it more powerful: dynamic changing between protocols
- Heartbeat (either in AYIYA or as a real Heartbeat) controls the user's IP (+port/proto)
- Incoming packets that match the details control wether we are in AYIYA, proto4/41 or GRE mode

This allows users to change tunnel type on _their_ end without informing the system
as the PoP can base this based on what it receives.

Note that there is no 'security issue' as heartbeats and AYIYA are authenticated on password
and switching between proto4/41 and GRE does not matter as they are 'insecure' anyway.